requirements.txt
sounddevice
numpy
openai
groq
keyboard
pywebview
